Progress!  Installed Ansible!
I was able to enable DNS using a couple of public Google DNS servers and the `--dns` docker option.  I thought I had tried something like this yesterday but it worked today.

I was also doing some stupid things in the Dockerfile just to try to resolve my earlier problems so the container wasn't even staying up when I first checked in the file but it does stay up now.