Providing myshell script
`myshell` is a simple shell that can replace `bash` in a sense but it's a super simple shell - the only commands it really understands is `date` (a very simple command) and `exit` (or CTRL-D!).  I basically wanted to see what Ansible would do when a user has a custom shell.  From superuser, you can use a command like the following to replace peon's shell:

	usermod --shell /usr/bin/myshell peon